requirements.txt
click
pillow
python-docx
